Dash board graph section fullview implementation (1248311222)

+++ app/src/main/java/com/perfect/prodsuit/Model/ScorePie.kt
@@ -1,4 +1,4 @@
 package com.perfect.prodsuit.Model
 
 data class ScorePie(val Piename:String,
-                    val Piescore: Int,)
+                    val Piescore: Float,)
